+++ integration_tests/s3/test_s3_deltalake.py
@@ -11,18 +11,44 @@ from deltalake import DeltaTable, write_deltalake
 import pathway as pw
 from pathway.tests.utils import CsvLinesNumberChecker, wait_result_with_checker
 
+from .base import MINIO_BUCKET_NAME, S3_BUCKET_NAME
 
-def write_deltalake_with_auth(s3_path, chunk, **kwargs):
+
+def get_deltalake_connection_options(storage_type):
+    options = {
+        "AWS_S3_ALLOW_UNSAFE_RENAME": "True",
+    }
+    if storage_type == "s3":
+        options.update(
+            {
+                "AWS_ACCESS_KEY_ID": os.environ["AWS_S3_ACCESS_KEY"],
+                "AWS_SECRET_ACCESS_KEY": os.environ["AWS_S3_SECRET_ACCESS_KEY"],
+                "AWS_REGION": "eu-central-1",
+                "AWS_BUCKET_NAME": S3_BUCKET_NAME,
+            }
+        )
+    elif storage_type == "minio":
+        options.update(
+            {
+                "AWS_ACCESS_KEY_ID": os.environ["MINIO_S3_ACCESS_KEY"],
+                "AWS_SECRET_ACCESS_KEY": os.environ["MINIO_S3_SECRET_ACCESS_KEY"],
+                "AWS_BUCKET_NAME": MINIO_BUCKET_NAME,
+                "AWS_ENDPOINT_URL": "https://minio-api.deploys.pathway.com",
+            }
+        )
+    else:
+        raise RuntimeError(
+            f"Unknown storage type: {storage_type}."
+            'Only "s3" and "minio" are supported.'
+        )
+    return options
+
+
+def write_deltalake_with_auth(storage_type, s3_path, chunk, **kwargs):
     write_deltalake(
         s3_path,
         chunk,
-        storage_options={
-            "AWS_ACCESS_KEY_ID": os.environ["AWS_S3_ACCESS_KEY"],
-            "AWS_SECRET_ACCESS_KEY": os.environ["AWS_S3_SECRET_ACCESS_KEY"],
-            "AWS_REGION": "eu-central-1",
-            "AWS_S3_ALLOW_UNSAFE_RENAME": "True",
-            "AWS_BUCKET_NAME": "aws-integrationtest",
-        },
+        storage_options=get_deltalake_connection_options(storage_type),
         **kwargs,
     )
 
@@ -33,14 +59,25 @@ def write_deltalake_with_auth(s3_path, chunk, **kwargs):
         pw.io.s3.AwsS3Settings(
             access_key=os.environ["AWS_S3_ACCESS_KEY"],
             secret_access_key=os.environ["AWS_S3_SECRET_ACCESS_KEY"],
-            bucket_name="aws-integrationtest",
+            bucket_name=S3_BUCKET_NAME,
             region="eu-central-1",
+        ),
+        pw.io.minio.MinIOSettings(
+            bucket_name=MINIO_BUCKET_NAME,
+            access_key=os.environ["MINIO_S3_ACCESS_KEY"],
+            secret_access_key=os.environ["MINIO_S3_SECRET_ACCESS_KEY"],
+            endpoint="minio-api.deploys.pathway.com",
         ),
         None,
     ],
 )
 def test_streaming_from_deltalake(credentials, tmp_path, s3_path):
-    lake_path = f"s3://aws-integrationtest/{s3_path}/"
+    if isinstance(credentials, pw.io.minio.MinIOSettings):
+        lake_path = f"s3://{MINIO_BUCKET_NAME}/{s3_path}/"
+        storage_type = "minio"
+    else:
+        lake_path = f"s3://{S3_BUCKET_NAME}/{s3_path}/"
+        storage_type = "s3"
     output_path = tmp_path / "output.csv"
 
     class InputSchema(pw.Schema):
@@ -49,13 +86,13 @@ def test_streaming_from_deltalake(credentials, tmp_path, s3_path):
 
     data = [{"k": 0, "v": ""}]
     df = pd.DataFrame(data).set_index("k")
-    write_deltalake_with_auth(lake_path, df, mode="append")
+    write_deltalake_with_auth(storage_type, lake_path, df, mode="append")
 
     def create_new_versions(start_idx, end_idx):
         for idx in range(start_idx, end_idx):
             data = [{"k": idx, "v": "a" * idx}]
             df = pd.DataFrame(data).set_index("k")
-            write_deltalake_with_auth(lake_path, df, mode="append")
+            write_deltalake_with_auth(storage_type, lake_path, df, mode="append")
             time.sleep(1.0)
 
     t = threading.Thread(target=create_new_versions, args=(1, 10))
@@ -76,8 +113,14 @@ def test_streaming_from_deltalake(credentials, tmp_path, s3_path):
         pw.io.s3.AwsS3Settings(
             access_key=os.environ["AWS_S3_ACCESS_KEY"],
             secret_access_key=os.environ["AWS_S3_SECRET_ACCESS_KEY"],
-            bucket_name="aws-integrationtest",
+            bucket_name=S3_BUCKET_NAME,
             region="eu-central-1",
+        ),
+        pw.io.minio.MinIOSettings(
+            bucket_name=MINIO_BUCKET_NAME,
+            access_key=os.environ["MINIO_S3_ACCESS_KEY"],
+            secret_access_key=os.environ["MINIO_S3_SECRET_ACCESS_KEY"],
+            endpoint="minio-api.deploys.pathway.com",
         ),
         None,
     ],
@@ -85,7 +128,12 @@ def test_streaming_from_deltalake(credentials, tmp_path, s3_path):
 @pytest.mark.parametrize("min_commit_frequency", [None, 60_000])
 def test_output(credentials, min_commit_frequency, tmp_path, s3_path):
     input_path = tmp_path / "input.csv"
-    output_s3_path = f"s3://aws-integrationtest/{s3_path}/"
+    if isinstance(credentials, pw.io.minio.MinIOSettings):
+        output_s3_path = f"s3://{MINIO_BUCKET_NAME}/{s3_path}/"
+        storage_type = "minio"
+    else:
+        output_s3_path = f"s3://{S3_BUCKET_NAME}/{s3_path}/"
+        storage_type = "s3"
     input_contents = "key,value\n1,Hello\n2,World"
 
     with open(input_path, "w") as f:
@@ -104,7 +152,9 @@ def test_output(credentials, min_commit_frequency, tmp_path, s3_path):
     )
     pw.run(monitoring_level=pw.MonitoringLevel.NONE)
 
-    delta_table = DeltaTable(output_s3_path)
+    delta_table = DeltaTable(
+        output_s3_path, storage_options=get_deltalake_connection_options(storage_type)
+    )
     pd_table_from_delta = delta_table.to_pandas()
     assert pd_table_from_delta.shape[0] == 2
 
@@ -115,19 +165,30 @@ def test_output(credentials, min_commit_frequency, tmp_path, s3_path):
         pw.io.s3.AwsS3Settings(
             access_key=os.environ["AWS_S3_ACCESS_KEY"],
             secret_access_key=os.environ["AWS_S3_SECRET_ACCESS_KEY"],
-            bucket_name="aws-integrationtest",
+            bucket_name=S3_BUCKET_NAME,
             region="eu-central-1",
+        ),
+        pw.io.minio.MinIOSettings(
+            bucket_name=MINIO_BUCKET_NAME,
+            access_key=os.environ["MINIO_S3_ACCESS_KEY"],
+            secret_access_key=os.environ["MINIO_S3_SECRET_ACCESS_KEY"],
+            endpoint="minio-api.deploys.pathway.com",
         ),
         None,
     ],
 )
 def test_input(credentials, tmp_path, s3_path):
-    input_s3_path = f"s3://aws-integrationtest/{s3_path}/"
+    if isinstance(credentials, pw.io.minio.MinIOSettings):
+        input_s3_path = f"s3://{MINIO_BUCKET_NAME}/{s3_path}/"
+        storage_type = "minio"
+    else:
+        input_s3_path = f"s3://{S3_BUCKET_NAME}/{s3_path}/"
+        storage_type = "s3"
     output_path = tmp_path / "output.csv"
 
     data = [{"k": 1, "v": "one"}, {"k": 2, "v": "two"}, {"k": 3, "v": "three"}]
     original = pd.DataFrame(data).set_index("k")
-    write_deltalake_with_auth(input_s3_path, original)
+    write_deltalake_with_auth(storage_type, input_s3_path, original)
 
     class InputSchema(pw.Schema):
         k: int = pw.column_definition(primary_key=True)